+++ indexer.go
@@ -6,7 +6,9 @@ package main
 
 import (
 	"context"
+	"time"
 
+	"github.com/elastic/package-registry/metrics"
 	"github.com/elastic/package-registry/packages"
 )
 
@@ -32,6 +34,8 @@ func (c CombinedIndexer) Init(ctx context.Context) error {
 }
 
 func (c CombinedIndexer) Get(ctx context.Context, opts *packages.GetOptions) (packages.Packages, error) {
+	start := time.Now()
+	defer metrics.StorageIndexerGetDurationSeconds.Observe(time.Since(start).Seconds())
 	var packages packages.Packages
 	for _, indexer := range c {
 		p, err := indexer.Get(ctx, opts)